refactor: extract EmojiPicker to plugins
