--- src/components/Emojis/index.ts
@@ -1,1 +1,0 @@
-export * from './EmojiPicker';
