+++ src/plugins/Emojis/index.ts
@@ -1,0 +1,2 @@
+export * from './EmojiPicker';
+export { EmojiPickerIcon } from './icons';
